@@ -1,60 +1,7 @@
 <?php
 
-namespace beacon;
+namespace beacon\core;
 
-/**
- * Created by PhpStorm.
- * User: wj008
- * Date: 2017/12/13
- * Time: 14:03
- */
-
-
-use \PDO as PDO;
-
-/**
- * sql 语句片段,用于更新插入时使用
- * Class SqlSection
- * @package beacon
- */
-class SqlRaw
-{
-    public $sql = null;
-    public $args = null;
-
-    public function __construct(string $sql, $args = null)
-    {
-        $this->sql = $sql;
-        $this->args = $args;
-    }
-
-    public function format()
-    {
-        return Mysql::format($this->sql, $this->args);
-    }
-}
-
-/**
- * 错误处理
- * Class MysqlException
- * @package beacon
- */
-class MysqlException extends \Exception
-{
-
-    protected $detail = '';
-
-    public function __construct(string $message = '', $detail = '', int $code = 0, \Throwable $previous = null)
-    {
-        $this->detail = $detail;
-        parent::__construct($message, $code, $previous);
-    }
-
-    public function getDetail()
-    {
-        return $this->detail;
-    }
-}
 
 /**
  * mysql 数据操作类
@@ -63,44 +10,12 @@ class MysqlException extends \Exception
  */
 class Mysql
 {
-    private static $instance = null;
-
-    /**
-     * 获取一个单例
-     * @return Mysql|null
-     * @throws MysqlException
-     */
-    public static function instance()
-    {
-        if (self::$instance == null) {
-            $host = Config::get('db.db_host', '127.0.0.1');
-            $port = Config::get('db.db_port', 3306);
-            $name = Config::get('db.db_name', '');
-            $user = Config::get('db.db_user', '');
-            $pass = Config::get('db.db_pwd', '');
-            $prefix = Config::get('db.db_prefix', 'sl_');
-            $charset = Config::get('db.db_charset', 'utf8');
-            $timeout = Config::get('db.timeout', 120);
-
-            try {
-                self::$instance = new Mysql($host, $port, $name, $user, $pass, $prefix, $charset, $timeout);
-            } catch (\PDOException $e) {
-                self::$instance = null;
-                throw new MysqlException($e->getMessage(), '', $e->getCode(), $e);
-            } catch (\Exception $e) {
-                self::$instance = null;
-                throw new MysqlException($e->getMessage(), '', $e->getCode(), $e);
-            }
-        }
-        return self::$instance;
-    }
-
     /**
      * 编码 sql 语句
      * @param $value
-     * @return false|int|string
+     * @return string|int|float
      */
-    public static function escape($value)
+    public static function escape($value): string|int|float
     {
         if ($value === null) {
             return 'NULL';
@@ -126,28 +41,18 @@ class Mysql
                 break;
         }
         $value = '\'' . preg_replace_callback('@[\0\b\t\n\r\x1a\"\'\\\\]@', function ($m) {
-                switch ($m[0]) {
-                    case '\0':
-                        return '\\0';
-                    case '\b':
-                        return '\\b';
-                    case '\t':
-                        return '\\t';
-                    case '\n':
-                        return '\\n';
-                    case '\r':
-                        return '\\r';
-                    case '\x1a':
-                        return '\\Z';
-                    case '"':
-                        return '\\"';
-                    case '\'':
-                        return '\\\'';
-                    case '\\':
-                        return '\\\\';
-                    default:
-                        return '';
-                }
+                return match ($m[0]) {
+                    '\0' => '\\0',
+                    '\b' => '\\b',
+                    '\t' => '\\t',
+                    '\n' => '\\n',
+                    '\r' => '\\r',
+                    '\x1a' => '\\Z',
+                    '"' => '\\"',
+                    '\'' => '\\\'',
+                    '\\' => '\\\\',
+                    default => '',
+                };
             }, $value) . '\'';
         return $value;
     }
@@ -155,12 +60,12 @@ class Mysql
     /**
      * 格式化 sql 语句
      * @param string $sql
-     * @param null $args
-     * @return null|string|string[]
+     * @param array|null $args
+     * @return string
      */
-    public static function format(string $sql, $args = null)
+    public static function format(string $sql, mixed $args = null): string
     {
-        if ($args == null) {
+        if ($args === null) {
             return $sql;
         }
         if (!is_array($args)) {
@@ -168,7 +73,7 @@ class Mysql
         }
         if (preg_match('@\?@', $sql)) {
             $index = 0;
-            $sql = preg_replace_callback('@\?@', function ($match) use (&$args, &$index) {
+            $sql = preg_replace_callback('@\?@', function () use (&$args, &$index) {
                 if (!isset($args[$index])) {
                     $index++;
                     return '?';
@@ -182,27 +87,16 @@ class Mysql
     }
 
 
-    /**
-     * 前缀
-     * @var string
-     */
-    private $prefix = '';
-    /**
-     * @var \PDO|null
-     */
-    private $pdo = null;
-    /**
-     * 事务计数器,防止多次调用
-     * @var int
-     */
-    private $transactionCounter = 0;
-    private $_lastSql = '';
+    protected \PDO $pdo;
+    protected int $transactionCounter = 0;
 
-    private $link = null;
-    private $user = null;
-    private $pass = null;
-    private $charset = 'utf8';
-    private $timeout = 120;
+    protected string $_lastSql = '';
+    protected string $link = '';
+    protected string $user = '';
+    protected string $pass = '';
+    protected string $charset = 'utf8';
+    protected string $prefix = '';
+    protected int $timeout = 120;
 
     /**
      * 构造函数
@@ -216,7 +110,7 @@ class Mysql
      * @param string $charset
      * @param int $timeout
      */
-    public function __construct($host, $port = 3306, $name = '', $user = '', $pass = '', $prefix = '', $charset = 'utf8', int $timeout = 120)
+    public function __construct(string $host, int $port = 3306, string $name = '', string $user = '', string $pass = '', string $prefix = '', string $charset = 'utf8', int $timeout = 120)
     {
         $this->prefix = $prefix;
         if (!empty($name)) {
@@ -229,27 +123,22 @@ class Mysql
         $this->pass = $pass;
         $this->charset = $charset;
         $this->timeout = $timeout;
-        try {
-            if ($timeout == 0) {
-                $this->pdo = new PDO($link, $user, $pass, [PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
-            } else {
-                $this->pdo = new PDO($link, $user, $pass, [PDO::ATTR_PERSISTENT => true, PDO::ATTR_TIMEOUT => $timeout, PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
-            }
-        } catch (\PDOException $exc) {
-            throw $exc;
+        if ($timeout == 0) {
+            $this->pdo = new \PDO($link, $user, $pass, [\PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
+        } else {
+            $this->pdo = new \PDO($link, $user, $pass, [\PDO::ATTR_PERSISTENT => true, \PDO::ATTR_TIMEOUT => $timeout, \PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
         }
     }
 
-    public function reconnection()
+    /**
+     *重连数据库
+     */
+    public function reConnection()
     {
-        try {
-            if ($this->timeout == 0) {
-                $this->pdo = new PDO($this->link, $this->user, $this->pass, [PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
-            } else {
-                $this->pdo = new PDO($this->link, $this->user, $this->pass, [PDO::ATTR_PERSISTENT => true, PDO::ATTR_TIMEOUT => $this->timeout, PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
-            }
-        } catch (\PDOException $exc) {
-            throw $exc;
+        if ($this->timeout == 0) {
+            $this->pdo = new \PDO($this->link, $this->user, $this->pass, [\PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
+        } else {
+            $this->pdo = new \PDO($this->link, $this->user, $this->pass, [\PDO::ATTR_PERSISTENT => true, \PDO::ATTR_TIMEOUT => $this->timeout, \PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES " . $this->charset]);
         }
     }
 
@@ -257,7 +146,7 @@ class Mysql
      * 开启事务
      * @return bool
      */
-    public function beginTransaction()
+    public function beginTransaction(): bool
     {
         if (!$this->transactionCounter++) {
             return $this->pdo->beginTransaction();
@@ -270,7 +159,7 @@ class Mysql
      * 是否在事务里面
      * @return bool
      */
-    public function inTransaction()
+    public function inTransaction(): bool
     {
         return $this->transactionCounter > 0;
     }
@@ -279,7 +168,7 @@ class Mysql
      * 提交事务
      * @return bool
      */
-    public function commit()
+    public function commit(): bool
     {
         if (!--$this->transactionCounter) {
             return $this->pdo->commit();
@@ -290,9 +179,9 @@ class Mysql
     /**
      * 回滚事务
      * @return bool
-     * @throws MysqlException
+     * @throws DBException
      */
-    public function rollBack()
+    public function rollBack(): bool
     {
         if (--$this->transactionCounter) {
             $this->exec('ROLLBACK TO trans' . ($this->transactionCounter + 1));
@@ -304,26 +193,31 @@ class Mysql
     /**
      * 执行sql
      * @param string $sql
-     * @return int
-     * @throws MysqlException
+     * @return int|false
+     * @throws DBException
      */
-    public function exec(string $sql)
+    public function exec(string $sql): int|false
     {
         try {
             $sql = str_replace('@pf_', $this->prefix, $sql);
-            $ret = $this->pdo->exec($sql);
-            return $ret;
+            return $this->pdo->exec($sql);
         } catch (\Exception $exception) {
-            throw new MysqlException($exception->getMessage(), $sql, $exception->getCode(), $exception);
+            $code = $exception->getCode();
+            if (!is_int($code)) {
+                $code = 0;
+            }
+            $exception = new DBException($exception->getMessage(), $code, $exception);
+            $exception->setDetail($sql);
+            throw $exception;
         }
     }
 
     /**
      * 获取最后的插入的id
-     * @param null $name
+     * @param string|null $name
      * @return string
      */
-    public function lastInsertId($name = null)
+    public function lastInsertId(?string $name = null): string
     {
         return $this->pdo->lastInsertId($name);
     }
@@ -332,7 +226,7 @@ class Mysql
      * 获取最后执行语句,需要开启 DEBUG_MYSQL_LOG
      * @return string
      */
-    public function lastSql()
+    public function lastSql(): string
     {
         return $this->_lastSql;
     }
@@ -340,11 +234,11 @@ class Mysql
     /**
      * 执行sql 语句
      * @param string $sql
-     * @param null $args
-     * @return bool|\PDOStatement
-     * @throws MysqlException
+     * @param mixed $args
+     * @return false|\PDOStatement
+     * @throws DBException
      */
-    public function execute(string $sql, $args = null)
+    public function execute(string $sql, mixed $args = null): false|\PDOStatement
     {
         $sql = str_replace('@pf_', $this->prefix, $sql);
         if ($args !== null && !is_array($args)) {
@@ -355,30 +249,13 @@ class Mysql
             $time = microtime(true);
         }
         try {
-            $retry = 0;
-            redo:
             $sth = $this->pdo->prepare($sql);
-            $ret = false;
-            try {
-                $ret = $sth->execute($args);
-            } catch (\Exception $exception) {
-                $ret = false;
-            }
-            if ($ret === FALSE) {
-                $err = $sth->errorInfo();
-                if (isset($err[1]) && $err[1] == 2006) {
-                    $this->reconnection();
-                    if ($retry == 0) {
-                        $retry = 1;
-                        goto redo;
-                    }
-                }
+            $ret = $sth->execute($args);
+            if ($ret === false) {
                 $this->_lastSql = Mysql::format($sql, $args);
-                if (isset($err[2])) {
-                    throw new MysqlException('execute sql statement error:' . $err[0] . ',' . $err[1] . ',' . $err[2], $this->_lastSql);
-                } else {
-                    throw new MysqlException('execute sql statement error:', $this->_lastSql);
-                }
+                $exception = new DBException('execute sql statement error.');
+                $exception->setDetail($this->_lastSql);
+                throw $exception;
             }
             if (defined('DEBUG_MYSQL_LOG') && DEBUG_MYSQL_LOG) {
                 $this->_lastSql = Mysql::format($sql, $args);
@@ -387,38 +264,28 @@ class Mysql
             return $sth;
         } catch (\Exception $exception) {
             $this->_lastSql = Mysql::format($sql, $args);
-            if (defined('DEBUG_MYSQL_LOG') && DEBUG_MYSQL_LOG) {
-                Logger::sql($this->_lastSql, microtime(true) - $time);
+            $code = $exception->getCode();
+            if (!is_int($code)) {
+                $code = 0;
             }
-            throw new MysqlException($exception->getMessage(), $this->_lastSql, $exception->getCode(), $exception);
+            $exception = new DBException($exception->getMessage(), $code, $exception);
+            $exception->setDetail($this->_lastSql);
+            throw $exception;
         }
     }
 
     /**
      * 获取多行记录
      * @param string $sql
-     * @param null $args
-     * @param null $fetch_style
-     * @param null $fetch_argument
-     * @param array|null $ctor_args
+     * @param ?array $args
+     * @param int $fetch_style
      * @return array
-     * @throws MysqlException
+     * @throws DBException
      */
-    public function getList(string $sql, $args = null, $fetch_style = null, $fetch_argument = null, array $ctor_args = null)
+    public function getList(string $sql, mixed $args = null, int $fetch_style = \PDO::FETCH_ASSOC): array
     {
-        if ($fetch_style === null) {
-            $fetch_style = PDO::FETCH_ASSOC;
-        }
         $stm = $this->execute($sql, $args);
-        if ($fetch_style !== null && $fetch_argument !== null && $ctor_args !== null) {
-            $rows = $stm->fetchAll($fetch_style, $fetch_argument, $ctor_args);
-        } elseif ($fetch_style !== null && $fetch_argument !== null) {
-            $rows = $stm->fetchAll($fetch_style, $fetch_argument);
-        } elseif ($fetch_style !== null) {
-            $rows = $stm->fetchAll($fetch_style);
-        } else {
-            $rows = $stm->fetchAll();
-        }
+        $rows = $stm->fetchAll($fetch_style);
         $stm->closeCursor();
         return $rows;
     }
@@ -426,36 +293,44 @@ class Mysql
     /**
      * 获取单行记录
      * @param string $sql
-     * @param null $args
-     * @param null $fetch_style
-     * @param null $cursor_orientation
-     * @param int $cursor_offset
-     * @return mixed|null
-     * @throws MysqlException
+     * @param array|null $args
+     * @param int $fetch_style
+     * @return mixed
+     * @throws DBException
      */
-    public function getRow(string $sql, $args = null, $fetch_style = null, $cursor_orientation = null, $cursor_offset = 0)
+    public function getRow(string $sql, mixed $args = null, int $fetch_style = \PDO::FETCH_ASSOC): mixed
     {
-        if ($fetch_style === null) {
-            $fetch_style = PDO::FETCH_ASSOC;
-        }
         $stm = $this->execute($sql, $args);
-        $row = $stm->fetch($fetch_style, $cursor_orientation, $cursor_offset);
+        $row = $stm->fetch($fetch_style);
         $stm->closeCursor();
         return $row === false ? null : $row;
     }
 
     /**
+     * 更加ID获取1行数据
+     * @param string $table
+     * @param int $id
+     * @return ?array
+     * @throws DBException
+     */
+    public function getItem(string $table, int $id): ?array
+    {
+        $sql = 'select * from `' . $table . '` where id=?';
+        return $this->getRow($sql, $id);
+    }
+
+    /**
      * 获得单个字段内容
      * @param string $sql
-     * @param null $args
-     * @param null $field
-     * @return mixed|null
-     * @throws MysqlException
+     * @param ?array $args
+     * @param ?string $field
+     * @return mixed
+     * @throws DBException
      */
-    public function getOne(string $sql, $args = null, $field = null)
+    public function getOne(string $sql, mixed $args = null, ?string $field = null): mixed
     {
         $row = $this->getRow($sql, $args);
-        if ($row == null) {
+        if ($row === null) {
             return null;
         }
         if (is_string($field) && !empty($field)) {
@@ -468,17 +343,17 @@ class Mysql
      * 获取某个字段的最大值
      * @param string $tbname
      * @param string $field
-     * @param null $where
-     * @param null $args
-     * @return null
-     * @throws MysqlException
+     * @param string|int|null $where
+     * @param mixed $args
+     * @return mixed
+     * @throws DBException
      */
-    public function getMax(string $tbname, string $field, $where = null, $args = null)
+    public function getMax(string $tbname, string $field, string|int|null $where = null, mixed $args = null): mixed
     {
         $sql = "select max(`{$field}`) from {$tbname}";
         if ($where !== null) {
             $where = trim($where);
-            if ($args != null) {
+            if ($args !== null) {
                 $args = is_array($args) ? $args : [$args];
             }
             if (is_int($where) || is_numeric($where)) {
@@ -487,8 +362,8 @@ class Mysql
             }
             $sql .= ' where ' . $where;
         }
-        $row = $this->getRow($sql, $args, PDO::FETCH_NUM);
-        if ($row == null) {
+        $row = $this->getRow($sql, $args, \PDO::FETCH_NUM);
+        if ($row === null) {
             return null;
         }
         return $row[0];
@@ -498,12 +373,12 @@ class Mysql
      * 获取某个字段的最小值
      * @param string $tbname
      * @param string $field
-     * @param null $where
-     * @param null $args
-     * @return null
-     * @throws MysqlException
+     * @param string|int|null $where
+     * @param mixed $args
+     * @return mixed
+     * @throws DBException
      */
-    public function getMin(string $tbname, string $field, $where = null, $args = null)
+    public function getMin(string $tbname, string $field, string|int|null $where = null, mixed $args = null): mixed
     {
         $sql = "select min(`{$field}`) from {$tbname}";
         if ($where !== null) {
@@ -517,7 +392,7 @@ class Mysql
             }
             $sql .= ' where ' . $where;
         }
-        $row = $this->getRow($sql, $args, PDO::FETCH_NUM);
+        $row = $this->getRow($sql, $args, \PDO::FETCH_NUM);
         if ($row == null) {
             return null;
         }
@@ -527,20 +402,19 @@ class Mysql
     /**
      * 创建一个sql语句原义片段,一般用于更新 插入数据时数组的值
      * @param string $sql
-     * @param null $args
+     * @param mixed $args
      * @return SqlRaw
      */
-    public function raw(string $sql, $args = null)
+    public function raw(string $sql, mixed $args = null): SqlRaw
     {
-        $data = new SqlRaw($sql, $args);
-        return $data;
+        return new SqlRaw($sql, $args);
     }
 
     /**
      * 插入记录
      * @param string $tbname
      * @param array $values
-     * @throws MysqlException
+     * @throws DBException
      */
     public function insert(string $tbname, array $values = [])
     {
@@ -548,14 +422,14 @@ class Mysql
             return;
         }
         $names = [];
-        $vals = [];
+        $params = [];
         $temp = [];
         foreach ($values as $key => $item) {
             $names[] = '`' . $key . '`';
             if ($item === null) {
-                $vals [] = 'NULL';
+                $params [] = 'NULL';
             } else if ($item instanceof SqlRaw) {
-                $vals [] = $item->sql;
+                $params [] = $item->sql;
                 if (is_array($item->args)) {
                     foreach ($item->args as $it) {
                         $temp[] = $it;
@@ -564,7 +438,7 @@ class Mysql
                     $temp[] = $item->args;
                 }
             } else {
-                $vals [] = '?';
+                $params [] = '?';
                 $type = gettype($item);
                 switch ($type) {
                     case 'bool':
@@ -572,12 +446,10 @@ class Mysql
                         $temp[] = $item ? 1 : 0;
                         break;
                     case 'int':
-                    case 'integer':
                     case 'double':
                     case 'float':
                     case 'string':
                         $temp[] = $item;
-                        break;
                         break;
                     case 'array':
                     case 'object':
@@ -589,7 +461,7 @@ class Mysql
                 }
             }
         }
-        $sql = 'insert into ' . $tbname . '(' . join(',', $names) . ') values (' . join(',', $vals) . ')';
+        $sql = 'insert into ' . $tbname . '(' . join(',', $names) . ') values (' . join(',', $params) . ')';
         $Stm = $this->execute($sql, $temp);
         $Stm->closeCursor();
     }
@@ -598,7 +470,7 @@ class Mysql
      * 替换记录集
      * @param string $tbname
      * @param array $values
-     * @throws MysqlException
+     * @throws DBException
      */
     public function replace(string $tbname, array $values = [])
     {
@@ -606,14 +478,14 @@ class Mysql
             return;
         }
         $names = [];
-        $vals = [];
+        $params = [];
         $temp = [];
         foreach ($values as $key => $item) {
             $names[] = '`' . $key . '`';
             if ($item === null) {
-                $vals [] = 'NULL';
+                $params [] = 'NULL';
             } else if ($item instanceof SqlRaw) {
-                $vals [] = $item->sql;
+                $params [] = $item->sql;
                 if (is_array($item->args)) {
                     foreach ($item->args as $it) {
                         $temp[] = $it;
@@ -622,7 +494,7 @@ class Mysql
                     $temp[] = $item->args;
                 }
             } else {
-                $vals [] = '?';
+                $params [] = '?';
                 $type = gettype($item);
                 switch ($type) {
                     case 'bool':
@@ -630,12 +502,10 @@ class Mysql
                         $temp[] = $item ? 1 : 0;
                         break;
                     case 'int':
-                    case 'integer':
                     case 'double':
                     case 'float':
                     case 'string':
                         $temp[] = $item;
-                        break;
                         break;
                     case 'array':
                     case 'object':
@@ -647,7 +517,7 @@ class Mysql
                 }
             }
         }
-        $sql = 'replace into ' . $tbname . '(' . join(',', $names) . ') values (' . join(',', $vals) . ')';
+        $sql = 'replace into ' . $tbname . '(' . join(',', $names) . ') values (' . join(',', $params) . ')';
         $Stm = $this->execute($sql, $temp);
         $Stm->closeCursor();
     }
@@ -656,11 +526,11 @@ class Mysql
      * 更新记录集
      * @param string $tbname
      * @param array $values
-     * @param null $where
-     * @param null $args
-     * @throws MysqlException
+     * @param string|int|null $where
+     * @param mixed $args
+     * @throws DBException
      */
-    public function update(string $tbname, array $values, $where = null, $args = null)
+    public function update(string $tbname, array $values, string|int|null $where = null, mixed $args = null)
     {
         if (count($values) == 0) {
             return;
@@ -693,12 +563,10 @@ class Mysql
                         $temp[] = $item ? 1 : 0;
                         break;
                     case 'int':
-                    case 'integer':
                     case 'double':
                     case 'float':
                     case 'string':
                         $temp[] = $item;
-                        break;
                         break;
                     case 'array':
                     case 'object':
@@ -712,7 +580,7 @@ class Mysql
         }
         $sql = 'update ' . $tbname . ' set ' . join(',', $maps);
         if (empty($where)) {
-            throw new MysqlException('编辑数据必须带有条件');
+            throw new DBException('编辑数据必须带有条件');
         }
         $sql .= ' where ' . $where;
         if (is_array($args)) {
@@ -729,20 +597,20 @@ class Mysql
     /**
      * 删除记录集
      * @param string $tbname
-     * @param null $where
-     * @param null $args
-     * @throws MysqlException
+     * @param string|int|null $where
+     * @param mixed $args
+     * @throws DBException
      */
-    public function delete(string $tbname, $where = null, $args = null)
+    public function delete(string $tbname, string|int|null $where = null, mixed $args = null)
     {
         $where = trim($where);
         if (is_int($where) || is_numeric($where)) {
             $args = [intval($where)];
             $where = 'id=?';
         }
-        $sql = 'DELETE FROM ' . $tbname;
+        $sql = 'delete from ' . $tbname;
         if (empty($where)) {
-            throw new MysqlException('删除数据必须带有条件');
+            throw new DBException('删除数据必须带有条件');
         }
         $sql .= ' where ' . $where;
         $Stm = $this->execute($sql, $args);
@@ -753,9 +621,9 @@ class Mysql
      * 获取表字段
      * @param string $tbname
      * @return array
-     * @throws MysqlException
+     * @throws DBException
      */
-    public function getFields(string $tbname)
+    public function getFields(string $tbname): array
     {
         return $this->getList('show full fields from `' . $tbname . '`');
     }
@@ -765,9 +633,9 @@ class Mysql
      * @param string $tbname
      * @param string $field
      * @return bool
-     * @throws MysqlException
+     * @throws DBException
      */
-    public function existsField(string $tbname, string $field)
+    public function existsField(string $tbname, string $field): bool
     {
         return $this->getRow('DESCRIBE `' . $tbname . '` `' . $field . '`;') !== null;
     }
@@ -776,7 +644,7 @@ class Mysql
      * 创建数据库表
      * @param string $tbname
      * @param array $options
-     * @throws MysqlException
+     * @throws DBException
      */
     public function createTable(string $tbname, array $options = [])
     {
@@ -786,12 +654,12 @@ class Mysql
             'comment' => '',
         ], $options);
         if ($this->existsTable($tbname)) {
-            throw new MysqlException("数据库表已经存在,{$tbname}");
+            throw new DBException("数据库表已经存在,{$tbname}");
         }
         $sql = "create table `{$tbname}` (`id` int(11) not null auto_increment,primary key (`id`)) engine={$options['engine']} default charset={$options['charset']} comment=?";
         $stm = $this->execute($sql, [$options['comment']]);
         if (!$stm) {
-            throw new MysqlException("创建数据库表失败,{$tbname}");
+            throw new DBException("创建数据库表失败,{$tbname}");
         }
     }
 
@@ -800,10 +668,10 @@ class Mysql
      * @param string $tbname
      * @param string $field
      * @param array $options
-     * @return int
-     * @throws MysqlException
+     * @return false|int
+     * @throws DBException
      */
-    public function addField(string $tbname, string $field, array $options = [])
+    public function addField(string $tbname, string $field, array $options = []): false|int
     {
         $options = array_merge([
             'type' => 'VARCHAR',
@@ -813,27 +681,14 @@ class Mysql
             'comment' => '',
         ], $options);
 
-        list($type, $len, $scale, $def, $comment) = array_values($options);
+        [$type, $len, $scale, $def, $comment] = array_values($options);
         $type = strtoupper($type);
         $sql = "ALTER TABLE {$tbname} ADD `{$field}`";
-        switch ($type) {
-            case 'VARCHAR':
-            case 'INT':
-            case 'BIGINT':
-            case 'SMALLINT':
-            case 'INTEGER':
-            case 'TINYINT':
-                $sql .= $type . '(' . $len . ')';
-                break;
-            case 'DECIMAL':
-            case 'DOUBLE':
-            case 'FLOAT':
-                $sql .= $type . '(' . $len . ',' . $scale . ')';
-                break;
-            default:
-                $sql .= $type;
-                break;
-        }
+        $sql .= match ($type) {
+            'VARCHAR', 'INT', 'BIGINT', 'SMALLINT', 'INTEGER', 'TINYINT' => $type . '(' . $len . ')',
+            'DECIMAL', 'DOUBLE', 'FLOAT' => $type . '(' . $len . ',' . $scale . ')',
+            default => $type,
+        };
         if (!in_array(strtoupper($type), ['BLOB', 'TEXT', 'GEOMETRY', 'JSON'])) {
             $sql .= ' DEFAULT ' . Mysql::escape($def);
         }
@@ -849,10 +704,10 @@ class Mysql
      * @param string $tbname
      * @param string $field
      * @param array $options
-     * @return int
-     * @throws MysqlException
+     * @return false|int
+     * @throws DBException
      */
-    public function modifyField(string $tbname, string $field, array $options = [])
+    public function modifyField(string $tbname, string $field, array $options = []): false|int
     {
         $chkNew = $this->existsField($tbname, $field);
         if (!$chkNew) {
@@ -865,27 +720,14 @@ class Mysql
             'def' => null,
             'comment' => '',
         ], $options);
-        list($type, $len, $scale, $def, $comment) = array_values($options);
+        [$type, $len, $scale, $def, $comment] = array_values($options);
         $type = strtoupper($type);
         $sql = "ALTER TABLE {$tbname} MODIFY `{$field}`";
-        switch ($type) {
-            case 'VARCHAR':
-            case 'INT':
-            case 'BIGINT':
-            case 'SMALLINT':
-            case 'INTEGER':
-            case 'TINYINT':
-                $sql .= $type . '(' . $len . ')';
-                break;
-            case 'DECIMAL':
-            case 'DOUBLE':
-            case 'FLOAT':
-                $sql .= $type . '(' . $len . ',' . $scale . ')';
-                break;
-            default:
-                $sql .= $type;
-                break;
-        }
+        $sql .= match ($type) {
+            'VARCHAR', 'INT', 'BIGINT', 'SMALLINT', 'INTEGER', 'TINYINT' => $type . '(' . $len . ')',
+            'DECIMAL', 'DOUBLE', 'FLOAT' => $type . '(' . $len . ',' . $scale . ')',
+            default => $type,
+        };
         if (!in_array(strtoupper($type), ['BLOB', 'TEXT', 'GEOMETRY', 'JSON'])) {
             $sql .= ' DEFAULT ' . Mysql::escape($def);
         }
@@ -899,26 +741,25 @@ class Mysql
     /**
      * 更新字段
      * @param string $tbname
-     * @param string $oldfield
-     * @param string $newfield
+     * @param string $oldField
+     * @param string $newField
      * @param array $options
-     * @return int
-     * @throws MysqlException
+     * @return false|int
+     * @throws DBException
      */
-    public function updateField(string $tbname, string $oldfield, string $newfield, array $options = [])
+    public function updateField(string $tbname, string $oldField, string $newField, array $options = []): false|int
     {
-        if ($oldfield == $newfield) {
-            return $this->modifyField($tbname, $newfield, $options);
+        if ($oldField == $newField) {
+            return $this->modifyField($tbname, $newField, $options);
         }
-        $chkNew = $this->existsField($tbname, $newfield);
+        $chkNew = $this->existsField($tbname, $newField);
         if ($chkNew) {
-            return $this->modifyField($tbname, $newfield, $options);
+            return $this->modifyField($tbname, $newField, $options);
         }
-        $chkOld = $this->existsField($tbname, $oldfield);
-        if (!$chkOld && !$chkNew) {
-            return $this->addField($tbname, $newfield, $options);
+        $chkOld = $this->existsField($tbname, $oldField);
+        if (!$chkOld) {
+            return $this->addField($tbname, $newField, $options);
         }
-
         $options = array_merge([
             'type' => 'VARCHAR',
             'len' => 250,
@@ -926,27 +767,14 @@ class Mysql
             'def' => null,
             'comment' => '',
         ], $options);
-        list($type, $len, $scale, $def, $comment) = array_values($options);
+        [$type, $len, $scale, $def, $comment] = array_values($options);
         $type = strtoupper($type);
-        $sql = "ALTER TABLE {$tbname} CHANGE `{$oldfield}` `{$newfield}`";
-        switch ($type) {
-            case 'VARCHAR':
-            case 'INT':
-            case 'BIGINT':
-            case 'SMALLINT':
-            case 'INTEGER':
-            case 'TINYINT':
-                $sql .= $type . '(' . $len . ')';
-                break;
-            case 'DECIMAL':
-            case 'DOUBLE':
-            case 'FLOAT':
-                $sql .= $type . '(' . $len . ',' . $scale . ')';
-                break;
-            default:
-                $sql .= $type;
-                break;
-        }
+        $sql = "ALTER TABLE {$tbname} CHANGE `{$oldField}` `{$newField}`";
+        $sql .= match ($type) {
+            'VARCHAR', 'INT', 'BIGINT', 'SMALLINT', 'INTEGER', 'TINYINT' => $type . '(' . $len . ')',
+            'DECIMAL', 'DOUBLE', 'FLOAT' => $type . '(' . $len . ',' . $scale . ')',
+            default => $type,
+        };
         if (!in_array(strtoupper($type), ['BLOB', 'TEXT', 'GEOMETRY', 'JSON'])) {
             $sql .= ' DEFAULT ' . Mysql::escape($def);
         }
@@ -961,25 +789,25 @@ class Mysql
      * 删除字段
      * @param string $tbname
      * @param string $field
-     * @return int|null
-     * @throws MysqlException
+     * @return false|int
+     * @throws DBException
      */
-    public function dropField(string $tbname, string $field)
+    public function dropField(string $tbname, string $field): false|int
     {
         if ($this->existsField($tbname, $field)) {
             $sql = "ALTER TABLE {$tbname} DROP `{$field}`;";
             return $this->exec($sql);
         }
-        return null;
+        return 0;
     }
 
     /**
      * 检查表是否存在
      * @param string $tbname
      * @return bool
-     * @throws MysqlException
+     * @throws DBException
      */
-    public function existsTable(string $tbname)
+    public function existsTable(string $tbname): bool
     {
         $tbname = str_replace('@pf_', $this->prefix, $tbname);
         $row = $this->getRow('SHOW TABLES LIKE ?;', $tbname);
@@ -989,10 +817,10 @@ class Mysql
     /**
      * 删除表
      * @param string $tbname
-     * @return int
-     * @throws MysqlException
+     * @return int|false
+     * @throws DBException
      */
-    public function dropTable(string $tbname)
+    public function dropTable(string $tbname): int|false
     {
         $tbname = str_replace('@pf_', $this->prefix, $tbname);
         return $this->exec('DROP TABLE IF EXISTS ' . Mysql::escape($tbname) . ';');
