@@ -1,57 +1,140 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: wj008
- * Date: 2017/12/20
- * Time: 0:40
- */
+
 
 namespace beacon\widget;
 
-use beacon\Field;
-use beacon\Form;
-use beacon\Request;
-use beacon\Utils;
-use beacon\View;
 
+use beacon\core\Field;
+use beacon\core\Form;
+use beacon\core\Logger;
+use beacon\core\Request;
+use beacon\core\Util;
+use beacon\core\Validator;
+use beacon\core\View;
 
-class Container extends Hidden
+#[\Attribute]
+class Container extends Field
 {
     /**
-     * 完善字段信息
-     * @param Field $field
-     * @param Form $form
+     * @var string 部件类名称
+     */
+    public string $itemClass = '';
+
+    /**
+     * @var string 每项类型
+     */
+    public string $itemType = 'array';
+
+    /**
+     * 子表单类型
+     * @var string
+     */
+    public string $subType = 'add';
+    /**
+     * 缓存设置数据
+     * @var array
+     */
+    protected array $cacheData = [];
+
+    /**
+     * 子表单错误
+     * @var array
+     */
+    public array $childError = [];
+    /**
+     * 用于修正表单的方法
+     * @var string|array|null
+     */
+    public string|array|null $modifyFunc = null;
+
+    /**
+     * 指定模板
+     * @var string
+     */
+    public string $template = '';
+
+    /**
+     * 指定模板
+     * @var bool
+     */
+    public bool $removeBtn = true;
+    /**
+     * 插入按钮
+     * @var bool
+     */
+    public bool $insertBtn = false;
+    /**
+     * 排序按钮
+     * @var bool
+     */
+    public bool $sortBtn = false;
+
+
+    /**
+     * Container constructor.
+     * @param array $args
+     */
+    public function setting(array $args)
+    {
+        parent::setting($args);
+        if (isset($args['itemClass']) && is_string($args['itemClass'])) {
+            $this->itemClass = $args['itemClass'];
+        }
+        if (isset($args['itemType']) && is_string($args['itemType'])) {
+            $this->itemType = $args['itemType'];
+        }
+        if (isset($args['subType']) && is_string($args['subType'])) {
+            $this->subType = $args['subType'];
+        }
+        if (isset($args['modifyFunc']) && is_callable($args['modifyFunc'])) {
+            $this->modifyFunc = $args['modifyFunc'];
+        }
+        if (isset($args['template']) && is_string($args['template'])) {
+            $this->template = $args['template'];
+        }
+        if (isset($args['removeBtn']) && is_bool($args['removeBtn'])) {
+            $this->removeBtn = $args['removeBtn'];
+        }
+        if (isset($args['insertBtn']) && is_bool($args['insertBtn'])) {
+            $this->insertBtn = $args['insertBtn'];
+        }
+        if (isset($args['sortBtn']) && is_bool($args['sortBtn'])) {
+            $this->sortBtn = $args['sortBtn'];
+        }
+    }
+
+    /**
+     * 完善表单嘻嘻
+     * @param Form $subForm
      * @param string $index
      */
-    public static function perfect(Field $field, Form $form, $index = '@@index@@')
+    private function perfectForm(Form $subForm, string $index = '@@index@@')
     {
-        $fields = $form->getFields();
+        $fields = $subForm->getFields();
         foreach ($fields as $name => $child) {
-            if ($field->dataValDisabled) {
-                $child->dataValDisabled = true;
+            if (isset($this->valid['disabled']) && $this->valid['disabled']) {
+                $child->valid['disabled'] = true;
             }
-            //当个
-            if ($field->mode == 'single') {
-                $child->boxId = $field->boxId . '_' . $child->boxId;
-                $child->boxName = $field->boxName . '[' . $child->boxName . ']';
-                $child->offEdit = $child->offEdit || $field->offEdit;
-                //如果存在拆分的时候
-                if ($child->names && is_array($child->names)) {
-                    $names = $child->names;
-                    foreach ($names as $nKey => $nItem) {
-                        if (is_array($nItem) && $nItem['field']) {
-                            $nItem['field'] = $field->boxName . '[' . $nItem['field'] . ']';
-                        } elseif (is_string($nItem)) {
-                            $nItem = $field->boxName . '[' . $nItem . ']';
-                        }
-                        $names[$nKey] = $nItem;
+            $boxId = $this->boxId;
+            $boxName = $this->boxName;
+            $child->setAttr('id', $boxId . '_' . $index . '_' . $child->boxId);
+            $child->setAttr('name', $boxName . '[' . $index . '][' . $child->boxName . ']');
+            $child->offEdit = $child->offEdit || $this->offEdit;
+            //如果存在拆分的时候
+            if (property_exists($child, 'names') && isset($child->names) && is_array($child->names)) {
+                $names = $child->names;
+                foreach ($names as $nKey => $name) {
+                    if (is_string($name)) {
+                        $names[$nKey] = $boxName . '[' . $index . '][' . $name . ']';
                     }
-                    $child->names = $names;
                 }
-                //修正动态数据
-                if (!empty($child->dynamic)) {
-                    $form->createDynamic($child);
-                    $dataDynamic = $child->dataDynamic;
+                $child->names = $names;
+            }
+            //修正动态数据
+            if (!empty($child->dynamic)) {
+                $child->createDynamic();
+                $dataDynamic = $child->getAttr('data-dynamic');
+                if (!empty($dataDynamic)) {
                     foreach ($dataDynamic as &$item) {
                         foreach (['show', 'hide', 'off', 'on'] as $key) {
                             if (isset($item[$key])) {
@@ -60,258 +143,241 @@ class Container extends Hidden
                                 }
                                 if (is_array($item[$key])) {
                                     foreach ($item[$key] as $idx => $xit) {
-                                        $item[$key][$idx] = $field->boxId . '_' . $xit;
+                                        $item[$key][$idx] = $boxId . '_' . $index . '_' . $xit;
                                     }
                                 }
                             }
                         }
                     }
-                    $child->dataDynamic = $dataDynamic;
                 }
-
-            } //多组
-            else if ($field->mode == 'multiple') {
-                $child->boxId = $field->boxId . '_' . $index . '_' . $child->boxId;
-                $child->boxName = $field->boxName . '[' . $index . '][' . $child->boxName . ']';
-                $child->offEdit = $child->offEdit || $field->offEdit;
-                //如果存在拆分的时候
-                if ($child->names && is_array($child->names)) {
-                    $names = $child->names;
-                    foreach ($names as $nKey => $nItem) {
-                        if (is_array($nItem) && $nItem['field']) {
-                            $nItem['field'] = $field->boxName . '[' . $index . '][' . $nItem['field'] . ']';
-                        } elseif (is_string($nItem)) {
-                            $nItem = $field->boxName . '[' . $index . '][' . $nItem . ']';
-                        }
-                        $names[$nKey] = $nItem;
-                    }
-                    $child->names = $names;
-                }
-                if (!empty($child->dynamic)) {
-                    $form->createDynamic($child);
-                    $dataDynamic = $child->dataDynamic;
-                    foreach ($dataDynamic as &$item) {
-                        foreach (['show', 'hide', 'off', 'on'] as $key) {
-                            if (isset($item[$key])) {
-                                if (is_string($item[$key])) {
-                                    $item[$key] = explode(',', $item[$key]);
-                                }
-                                if (is_array($item[$key])) {
-                                    foreach ($item[$key] as $idx => $xit) {
-                                        $item[$key][$idx] = $field->boxId . '_' . $index . '_' . $xit;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    $child->dataDynamic = $dataDynamic;
+                $child->setAttr('data-dynamic', $dataDynamic);
+            }
+            if (isset($this->valid['merge']) && $this->valid['merge'] === true && empty($child->valid['display'])) {
+                if (!empty($this->valid['display'])) {
+                    $child->valid['display'] = $this->valid['display'];
+                } else {
+                    $child->valid['display'] = '#' . $boxId . '-validation';
                 }
             }
-            $child->dataValOutput = '#' . $field->boxId . '-validation';
         }
     }
 
     /**
-     * 获取插件表单
-     * @param Field $field
-     * @param string|null $type
-     * @return Form
+     * 获取类名
+     * @return string
+     */
+    private function getClassName(): string
+    {
+        if (!empty($this->itemClass) && class_exists($this->itemClass)) {
+            return $this->itemClass;
+        }
+        $typeName = $this->itemType;
+        if ($typeName == 'array') {
+            return '';
+        }
+        if (!empty($typeName) && class_exists($typeName)) {
+            return $typeName;
+        }
+        return '';
+    }
+
+    /**
+     * 获取插件
+     * @param string|object $object
+     * @param string $type
+     * @return Form|null
      * @throws \Exception
      */
-    public static function plugForm(Field $field, string $type = null)
+    private function getSubForm(string|object $object, string $type = ''): ?Form
     {
-        $className = $field->plugName;
-        if (empty($className) || !class_exists($className)) {
-            throw new \Exception('没有找到插件的类 ' . $className);
+        if (empty($type)) {
+            $form = $this->getForm();
+            if (empty($form)) {
+                $type = $this->subType;
+            } else {
+                $type = $form->getType();
+            }
         }
-        if ($type) {
-            return new $className($type);
+        $subForm = Form::create($object, $type);
+        if ($subForm == null) {
+            throw new \Exception('子表单获取失败');
         }
-        return new $className($field->getForm()->getType());
+        if (!empty($this->modifyFunc) && is_callable($this->modifyFunc)) {
+            call_user_func($this->modifyFunc, $subForm);
+        }
+        return $subForm;
     }
 
     /**
-     * 生成代码
-     * @param Field $field
-     * @param array $attr
+     * @param array $attrs
      * @return string
      * @throws \Exception
      */
-    public function code(Field $field, $attr = [])
+    protected function code(array $attrs = []): string
     {
-        $field->mode = empty($field->mode) ? 'single' : $field->mode;
-        #单一模式
-        if ($field->mode == 'single') {
-            $attr['type'] = '';
-            $attr['name'] = '';
-            $attr['value'] = '';
-            $attr['class'] = '';
-            $attr['id'] = 'row_' . $field->boxId;
-            $attr = WidgetHelper::mergeAttributes($field, $attr);
-            if (empty($field->plugName)) {
-                return '<div ' . join(' ', $attr) . '></div>';
-            }
-            $form = self::plugForm($field);
-            if (empty($form->template)) {
-                throw new \Exception('插件模板不存在');
-            }
-            $viewer = new View();
-            $viewer->fetch($form->template);
-            $wrapFunc = $viewer->getHook('single');
-            if ($wrapFunc == null) {
-                throw new \Exception('模板中没有找到 {hook fn="single"} 的钩子函数');
-            }
-            $tinker = $field->getFunc('tinker');
-            if ($tinker && is_callable($tinker)) {
-                call_user_func($tinker, $form);
-            }
-            $form->setValues($field->value);
-            self::perfect($field, $form);
-            $code = $wrapFunc(['field' => $field, 'form' => $form]);
-            return '<div ' . join(' ', $attr) . '>' . $code . '</div>';
-        } else {
-            #多增模式
-            if (empty($field->plugName)) {
-                return '';
-            }
-            $form = self::plugForm($field, 'add');
-            if (empty($form->template)) {
-                throw new \Exception('插件模板不存在');
-            }
-            $viewer = new View();
-            $viewer->fetch($form->template);
-            $wrapFunc = $viewer->getHook('multiple-wrap');
-            $itemFunc = $viewer->getHook('multiple-item');
-            if ($wrapFunc == null) {
-                throw new \Exception('模板中没有找到 {hook fn="multiple-wrap"} 的钩子函数');
-            }
-            if ($itemFunc == null) {
-                throw new \Exception('模板中没有找到 {hook fn="multiple-item"} 的钩子函数');
-            }
-            $out = [];
-            $values = $field->value;
-            $index = 0;
-            $tinker = $field->getFunc('tinker');
-            if (!empty($values) && is_array($values)) {
-                foreach ($values as $idx => $item) {
-                    $pForm = self::plugForm($field, 'edit');
-                    if ($tinker && is_callable($tinker)) {
-                        call_user_func($tinker, $pForm);
-                    }
-                    $pForm->setValues($item);
-                    self::perfect($field, $pForm, $index);
-                    $code = $itemFunc(['field' => $field, 'form' => $pForm, 'index' => 'a' . $index]);
-                    $out[] = $code;
-                    $index++;
-                }
-            }
-            if ($tinker && is_callable($tinker)) {
-                call_user_func($tinker, $form);
-            }
-            self::perfect($field, $form);
-            $code = $itemFunc(['field' => $field, 'form' => $form, 'index' => '@@index@@']);
-            $data = [];
-            $data[] = 'data-index="' . htmlspecialchars($index) . '"';
-            if ($field->dataMinSize) {
-                $data[] = 'data-min-size="' . $field->dataMinSize . '"';
-            }
-            if ($field->dataMaxSize) {
-                $data[] = 'data-max-size="' . $field->dataMaxSize . '"';
-            }
-            if ($field->dataInitSize) {
-                $data[] = 'data-init-size="' . $field->dataInitSize . '"';
-            }
-            return $wrapFunc([
-                'field' => $field,
-                'form' => $form,
-                'body' => join('', $out),
-                'source' => base64_encode($code),
-                'lastIndex' => $index,
-                'attrs' => join(' ', $data)
-            ]);
+        $className = $this->getClassName();
+        if (empty($className)) {
+            return '';
         }
+        $subForm = $this->getSubForm($className, 'add');
+        if (empty($subForm->template) && empty($this->template)) {
+            throw new \Exception('子表单获模板未定义');
+        }
+
+        $viewer = new View();
+        $template = $subForm->template;
+        if (!empty($this->template)) {
+            $template = $this->template;
+        }
+        $viewer->fetch($template);
+        $wrapFunc = $viewer->getHook('wrap');
+        $itemFunc = $viewer->getHook('item');
+        if ($wrapFunc == null) {
+            throw new \Exception('模板中没有找到 {hook fn="wrap"} 的钩子函数');
+        }
+        if ($itemFunc == null) {
+            throw new \Exception('模板中没有找到 {hook fn="item"} 的钩子函数');
+        }
+        $code = [];
+        $index = 0;
+        $values = $this->cacheData;
+        if (!empty($values) && is_array($values)) {
+            foreach ($values as $item) {
+                $editForm = $this->getSubForm($className, 'add');
+                $editForm->setData($item);
+                $this->perfectForm($editForm, $index);
+                $subCode = '<div class="container-item">' . $itemFunc(['field' => $this, 'form' => $editForm, 'index' => 'a' . $index]) . '</div>';
+                $code[] = $subCode;
+                $index++;
+            }
+        }
+        $this->perfectForm($subForm);
+        $source = '<div class="container-item">' . $itemFunc(['field' => $this, 'form' => $subForm, 'index' => '@@index@@']) . '</div>';
+        $attrs['yee-module'] = $this->getYeeModule('container');
+        $attrs['data-index'] = $index;
+        $attrs['data-source'] = base64_encode($source);
+        $wrapStyle = '';
+        if (!empty($attrs['wrap-style'])) {
+            $wrapStyle = ' style="' . $attrs['wrap-style'] . '"';
+            unset($attrs['wrap-style']);
+        }
+        $data = [];
+        $data['field'] = $this;
+        $data['body'] = '<div class="container-wrap"' . $wrapStyle . '>' . join('', $code) . '</div>';
+        return static::makeTag('div', ['attrs' => $attrs, 'exclude' => ['name'], 'code' => $wrapFunc($data)]);
     }
 
     /**
-     * @param Field $field
-     * @param array $input
-     * @return array|null
+     * @param array $param
+     * @return mixed
      * @throws \Exception
      */
-    public function assign(Field $field, array $input)
+    public function fromParam(array $param = []): array
     {
-        $field->mode = empty($field->mode) ? 'single' : $field->mode;
-        $boxName = $field->boxName;
-        $itemData = Request::input($input, $boxName . ':a', $field->default);
-        if (empty($field->plugName)) {
-            return $field->value;
+        $boxName = $this->boxName;
+        $itemData = Request::lookType($param, $boxName, 'array', []);
+        $className = $this->getClassName();
+        if (empty($className)) {
+            return $itemData;
         }
-        //单个
-        if ($field->mode == 'single') {
-            $form = self::plugForm($field);
-            $tinker = $field->getFunc('tinker');
-            if ($tinker && is_callable($tinker)) {
-                call_user_func($tinker, $form);
+        $this->childError = [];
+        $this->cacheData = [];
+        $objects = [];
+        foreach ($itemData as $idx => $datum) {
+            if (!is_array($datum)) {
+                continue;
             }
-            $childValue = $form->fillComplete($itemData);
-            if ($field->dataValDisabled !== true) {
-                if (!$form->validation($errors)) {
-                    $childError = [];
-                    foreach ($errors as $key => $err) {
-                        $boxName = $field->boxName . '[' . $key . ']';
-                        $childError[$boxName] = $err;
-                    }
-                    $field->childError = $childError;
-                }
-            }
-            $field->value = $childValue;
-            return $field->value;
-        }
-        //多个
-        if ($field->mode == 'multiple') {
-            $temp = [];
-            $childError = [];
-            foreach ($itemData as $idx => $item) {
-                $form = self::plugForm($field);
-                $tinker = $field->getFunc('tinker');
-                if ($tinker && is_callable($tinker)) {
-                    call_user_func($tinker, $form);
-                }
-                $childValue = $form->fillComplete($item);
-                if ($field->dataValDisabled !== true) {
-                    $errors = [];
-                    if (!$form->validation($errors)) {
-                        foreach ($errors as $key => $err) {
-                            $boxName = $field->boxName . '[' . $idx . '][' . $key . ']';
-                            $childError[$boxName] = $err;
-                        }
+            $object = new $className();
+            $subForm = $this->getSubForm($object);
+            $subForm->fillComplete($datum);
+            $validDisabled = isset($this->valid['disabled']) && $this->valid['disabled'];
+            if (!$validDisabled) {
+                if (!$subForm->validate($errors)) {
+                    foreach ($errors as $key => $error) {
+                        $childName = $boxName . '[' . $idx . '][' . $key . ']';
+                        $this->childError[$childName] = $error;
                     }
                 }
-                $temp[] = $childValue;
             }
-            $field->childError = $childError;
-            $field->value = $temp;
-            return $field->value;
+            $objects[] = $object;
+            $this->cacheData[] = $subForm->getData();
         }
-    }
-
-    public function fill(Field $field, array &$values)
-    {
-        $values[$field->name] = json_encode($field->value, JSON_UNESCAPED_UNICODE);
-    }
-
-    public function init(Field $field, array $values)
-    {
-        $temp = isset($values[$field->name]) ? $values[$field->name] : null;
-        $value = null;
-        if (is_array($temp)) {
-            $value = $temp;
-        } else if (is_string($temp) && Utils::isJson($temp)) {
-            $value = json_decode($temp, true);
+        if ($this->itemType == 'array') {
+            return $this->cacheData;
         }
-        $field->value = $value;
+        return $objects;
     }
 
+
+    /**
+     * @param array $data
+     * @return array
+     * @throws \Exception
+     */
+    public function setData(array $data): array
+    {
+        $this->cacheData = $data;
+        if ($this->itemType == 'array') {
+            return $this->cacheData;
+        }
+        $className = $this->getClassName();
+        if (empty($className)) {
+            return [];
+        }
+        $objects = [];
+        foreach ($data as $datum) {
+            $object = new $className();
+            $subForm = $this->getSubForm($object);
+            $subForm->setData($datum);
+            $objects[] = $object;
+        }
+        return $objects;
+    }
+
+    /**
+     * 获取数据值
+     * @return array
+     */
+    public function getData(): array
+    {
+        return $this->cacheData;
+    }
+
+    /**
+     * @param array $data
+     * @return array
+     * @throws \Exception
+     */
+    public function fromData(array $data = []): array
+    {
+        $itemData = $data[$this->name] ?? null;
+        if (is_string($itemData) && Util::isJson($itemData)) {
+            $itemData = json_decode($itemData, true);
+        }
+        if (!is_array($itemData)) {
+            $itemData = [];
+        }
+        return $this->setData($itemData);
+    }
+
+
+    /**
+     * 验证数据
+     * @param array $errors
+     * @return bool
+     */
+    public function validate(array &$errors): bool
+    {
+        foreach ($this->childError as $childName => $error) {
+            if (!empty($error)) {
+                $errors[$childName] = $error;
+                if (empty($field->error)) {
+                    $this->error = $error;
+                }
+                return false;
+            }
+        }
+        return parent::validate($errors);
+    }
 
 }
-
